internal/envoy: move v2/route helpers into internal/envoy
Updates #708

Move route related helpers into internal/envoy.
