+++ internal/envoy/listener.go
@@ -11,9 +11,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package envoy contains a configuration writer for v2 YAML config.
-// To avoid a dependncy on a YAML library, we generate the YAML using
-// the text/template package.
 package envoy
 
 import (
